Extracted CommandArgs classes into their own files
