+++ src/Octoshift/Commands/WaitForMigration/WaitForMigrationCommandArgs.cs
@@ -1,0 +1,9 @@
+﻿
+namespace OctoshiftCLI.Commands.WaitForMigration;
+
+public class WaitForMigrationCommandArgs
+{
+    public string MigrationId { get; set; }
+    public string GithubPat { get; set; }
+    public bool Verbose { get; set; }
+}
