+++ src/Octoshift/Commands/WaitForMigration/WaitForMigrationCommandBase.cs
@@ -55,10 +55,3 @@ public class WaitForMigrationCommandBase : CommandBase<WaitForMigrationCommandAr
         return new WaitForMigrationCommandHandler(log, githubApi);
     }
 }
-
-public class WaitForMigrationCommandArgs
-{
-    public string MigrationId { get; set; }
-    public string GithubPat { get; set; }
-    public bool Verbose { get; set; }
-}
