+++ src/ado2gh/Commands/AddTeamToRepo/AddTeamToRepoCommand.cs
@@ -62,14 +62,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.AddTeamToRepo
             return new AddTeamToRepoCommandHandler(log, githubApi);
         }
     }
-
-    public class AddTeamToRepoCommandArgs
-    {
-        public string GithubOrg { get; set; }
-        public string GithubRepo { get; set; }
-        public string Team { get; set; }
-        public string Role { get; set; }
-        public string GithubPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
