+++ src/ado2gh/Commands/ConfigureAutoLink/ConfigureAutoLinkCommand.cs
@@ -61,14 +61,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.ConfigureAutoLink
             return new ConfigureAutoLinkCommandHandler(log, githubApi);
         }
     }
-
-    public class ConfigureAutoLinkCommandArgs
-    {
-        public string GithubOrg { get; set; }
-        public string GithubRepo { get; set; }
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string GithubPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
