+++ src/ado2gh/Commands/DisableRepo/DisableRepoCommand.cs
@@ -56,13 +56,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.DisableRepo
             return new DisableRepoCommandHandler(log, adoApi);
         }
     }
-
-    public class DisableRepoCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string AdoRepo { get; set; }
-        public string AdoPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
