+++ src/ado2gh/Commands/GenerateScript/GenerateScriptCommandArgs.cs
@@ -1,0 +1,24 @@
+﻿using System.IO;
+
+namespace OctoshiftCLI.AdoToGithub.Commands.GenerateScript
+{
+    public class GenerateScriptCommandArgs
+    {
+        public string GithubOrg { get; set; }
+        public string AdoOrg { get; set; }
+        public string AdoTeamProject { get; set; }
+        public FileInfo Output { get; set; }
+        public bool Sequential { get; set; }
+        public string AdoPat { get; set; }
+        public bool Verbose { get; set; }
+        public bool DownloadMigrationLogs { get; set; }
+        public bool CreateTeams { get; set; }
+        public bool LinkIdpGroups { get; set; }
+        public bool LockAdoRepos { get; set; }
+        public bool DisableAdoRepos { get; set; }
+        public bool IntegrateBoards { get; set; }
+        public bool RewirePipelines { get; set; }
+        public bool All { get; set; }
+        public FileInfo RepoList { get; set; }
+    }
+}
