+++ src/ado2gh/Commands/IntegrateBoards/IntegrateBoardsCommand.cs
@@ -64,15 +64,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.IntegrateBoards
             return new IntegrateBoardsCommandHandler(log, adoApi, environmentVariableProvider);
         }
     }
-
-    public class IntegrateBoardsCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string GithubOrg { get; set; }
-        public string GithubRepo { get; set; }
-        public string AdoPat { get; set; }
-        public string GithubPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
