+++ src/ado2gh/Commands/InventoryReport/InventoryReportCommand.cs
@@ -63,12 +63,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.InventoryReport
                 pipelinesCsvGeneratorService);
         }
     }
-
-    public class InventoryReportCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoPat { get; set; }
-        public bool Minimal { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
