+++ src/ado2gh/Commands/LockRepo/LockRepoCommand.cs
@@ -56,13 +56,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.LockRepo
             return new LockRepoCommandHandler(log, adoApi);
         }
     }
-
-    public class LockRepoCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string AdoRepo { get; set; }
-        public string AdoPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
