+++ src/ado2gh/Commands/LockRepo/LockRepoCommandArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace OctoshiftCLI.AdoToGithub.Commands.LockRepo
+{
+    public class LockRepoCommandArgs
+    {
+        public string AdoOrg { get; set; }
+        public string AdoTeamProject { get; set; }
+        public string AdoRepo { get; set; }
+        public string AdoPat { get; set; }
+        public bool Verbose { get; set; }
+    }
+}
