+++ src/ado2gh/Commands/ShareServiceConnection/ShareServiceConnectionCommand.cs
@@ -56,13 +56,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands.ShareServiceConnection
             return new ShareServiceConnectionCommandHandler(log, adoApi);
         }
     }
-
-    public class ShareServiceConnectionCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string ServiceConnectionId { get; set; }
-        public string AdoPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
