+++ src/ado2gh/Commands/ShareServiceConnection/ShareServiceConnectionCommandArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace OctoshiftCLI.AdoToGithub.Commands.ShareServiceConnection
+{
+    public class ShareServiceConnectionCommandArgs
+    {
+        public string AdoOrg { get; set; }
+        public string AdoTeamProject { get; set; }
+        public string ServiceConnectionId { get; set; }
+        public string AdoPat { get; set; }
+        public bool Verbose { get; set; }
+    }
+}
