+++ src/gei/Commands/MigrateCodeScanningAlerts/MigrateCodeScanningAlertsCommand.cs
@@ -94,18 +94,3 @@ public class MigrateCodeScanningAlertsCommand : CommandBase<MigrateCodeScanningA
         return new MigrateCodeScanningAlertsCommandHandler(log, codeScanningAlertService);
     }
 }
-
-public class MigrateCodeScanningAlertsCommandArgs
-{
-    public string SourceOrg { get; set; }
-    public string SourceRepo { get; set; }
-    public string TargetOrg { get; set; }
-    public string TargetRepo { get; set; }
-    public string TargetApiUrl { get; set; }
-    public string GhesApiUrl { get; set; }
-    public bool NoSslVerify { get; set; }
-    public bool Verbose { get; set; }
-    public bool DryRun { get; set; }
-    public string GithubSourcePat { get; set; }
-    public string GithubTargetPat { get; set; }
-}
