+++ src/gei/Commands/MigrateOrg/MigrateOrgCommand.cs
@@ -73,16 +73,4 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands.MigrateOrg
             return new MigrateOrgCommandHandler(log, targetGithubApi, environmentVariableProvider);
         }
     }
-
-    public class MigrateOrgCommandArgs
-    {
-        public string GithubSourceOrg { get; set; }
-        public string GithubTargetOrg { get; set; }
-        public string GithubTargetEnterprise { get; set; }
-        public bool Wait { get; set; }
-        public bool QueueOnly { get; set; }
-        public bool Verbose { get; set; }
-        public string GithubSourcePat { get; set; }
-        public string GithubTargetPat { get; set; }
-    }
 }
