+++ src/gei/Commands/MigrateRepo/MigrateRepoCommandArgs.cs
@@ -1,0 +1,34 @@
+﻿namespace OctoshiftCLI.GithubEnterpriseImporter.Commands.MigrateRepo
+{
+    public class MigrateRepoCommandArgs
+    {
+        public string GithubSourceOrg { get; set; }
+        public string AdoServerUrl { get; set; }
+        public string AdoSourceOrg { get; set; }
+        public string AdoTeamProject { get; set; }
+        public string SourceRepo { get; set; }
+        public string GithubTargetOrg { get; set; }
+        public string TargetRepo { get; set; }
+        public string TargetApiUrl { get; set; }
+        public string GhesApiUrl { get; set; }
+        public string AzureStorageConnectionString { get; set; }
+        public string AwsBucketName { get; set; }
+        public string AwsAccessKey { get; set; }
+        public string AwsSecretKey { get; set; }
+        public string AwsSessionToken { get; set; }
+        public string AwsRegion { get; set; }
+        public bool NoSslVerify { get; set; }
+        public string GitArchiveUrl { get; set; }
+        public string MetadataArchiveUrl { get; set; }
+        public bool SkipReleases { get; set; }
+        public bool LockSourceRepo { get; set; }
+        public bool Wait { get; set; }
+        public bool QueueOnly { get; set; }
+        public string TargetRepoVisibility { get; set; }
+        public bool Verbose { get; set; }
+        public string GithubSourcePat { get; set; }
+        public string GithubTargetPat { get; set; }
+        public string AdoPat { get; set; }
+        public bool KeepArchive { get; set; }
+    }
+}
